refactor: move MEDIA_TYPE to app-constants
