+++ src/Components/ChatComponent/components/ChatListItem/ChatListItem.jsx
@@ -9,9 +9,8 @@ import { FiLink2 } from "react-icons/fi";
 import { TbPhoto } from "react-icons/tb";
 import { GrAttachment } from "react-icons/gr";
 import { useApp } from "../../../../hooks";
-import { DEFAULT_PHOTO, HH_mm } from "../../../../app-constants";
+import { DEFAULT_PHOTO, HH_mm, MEDIA_TYPE } from "../../../../app-constants";
 import { Tag } from "../../../Tag";
-import { MEDIA_TYPE } from "../../renderContent";
 import {
   priorityTagColors,
   parseServerDate,
